renaming css & js to styles & scripts
