--- src/fonts/fonts.scss
@@ -1,17 +1,0 @@
-@font-face {
-  font-family: 'PT Sans';
-  src:
-    url('fonts/PTSans-Regular.woff2') format('woff2'),
-    url('fonts/PTSans-Regular.woff') format('woff');
-  font-weight: normal;
-  font-style: normal;
-}
-
-@font-face {
-  font-family: 'PT Serif';
-  src:
-    url('fonts/PTSerif-Regular.woff2') format('woff2'),
-    url('fonts/PTSerif-Regular.woff') format('woff');
-  font-weight: normal;
-  font-style: normal;
-}
