+++ src/styles/footer.scss
@@ -1,4 +1,4 @@
-@import "css/link";
+@import "styles/link";
 
 .footer {
   margin-top: 2.5rem;
