+++ src/styles/repo-details.scss
