+++ src/styles/rich-text.scss
@@ -1,5 +1,4 @@
-@import 'css/_variables';
-@import 'css/heading';
+@import 'styles/heading';
 
 .rich-text {
   h1,
